+++ app/controllers/facilities_management/rm6378/admin/uploads_controller.rb
@@ -6,6 +6,10 @@ module FacilitiesManagement
 
         private
 
+        def service_key
+          :fm
+        end
+
         def upload_params
           params.expect(facilities_management_rm6378_admin_upload: %i[supplier_details_file supplier_services_file supplier_regions_file]) if params[:facilities_management_rm6378_admin_upload].present?
         end